La primera parte de la ventana del portero. Ya se ven los eventos del dia actual y eventos del proximo dia

+++ src/com/itla/data/InvitadoAcessoDatos.java
@@ -21,7 +21,7 @@ public class InvitadoAcessoDatos {
         return invitados;
     }
 
-    public static Invitado SeleccionarPorId(int id) throws SQLException {
+    public static Invitado seleccionarPorId(int id) throws SQLException {
         Conexion.conectar();
         PreparedStatement ps = Conexion.conn.prepareStatement("SELECT * FROM evento WHERE ID_EVENTO = ?");
         ps.setInt(1, id);
@@ -39,42 +39,53 @@ public class InvitadoAcessoDatos {
             modificar(invitado);
             return;
         }
-
-        Conexion.conectar();
-        PreparedStatement ps = Conexion.conn.prepareStatement("insert into invitado values(SEC_ID_INVITADO.nextval, ?,?,?,?,?,?)");
-        ps.setString(1, invitado.getNombre());
-        ps.setString(2, invitado.getApellido());
-        ps.setString(3, invitado.getTelefono());
-        ps.setString(4, invitado.getDireccion());
-        ps.setString(5, String.valueOf(Conexion.convertirBooleanAChar(invitado.isActivo())));
-        ps.setString(6, invitado.getSexo());
-        ps.executeUpdate();
-        Conexion.desconectar();
+        try {
+            Conexion.conectar();
+            PreparedStatement ps = Conexion.conn.prepareStatement("insert into invitado values(SEC_ID_INVITADO.nextval, ?,?,?,?,?,?)");
+            ps.setString(1, invitado.getNombre());
+            ps.setString(2, invitado.getApellido());
+            ps.setString(3, invitado.getTelefono());
+            ps.setString(4, invitado.getDireccion());
+            ps.setString(5, String.valueOf(Conexion.convertirBooleanAChar(invitado.isActivo())));
+            ps.setString(6, invitado.getSexo());
+            ps.executeUpdate();
+        } finally {
+            Conexion.desconectar();
+        }
     }
 
     public static void modificar(Invitado invt) throws SQLException {
         if (invt.getId() <= 0) {
-            throw new IllegalArgumentException("No se puede modificar un registro con id 0");
+            throw new IllegalArgumentException("No se puede modificar un registro con id <= 0");
         }
-        Conexion.conectar();
-        PreparedStatement ps = Conexion.conn.prepareStatement("UPDATE evento SET NOMBRE = ?, FECHA = ?, UBICACION = ?, ACTIVO = ? WHERE ID_EVENTO = ?");
-        ps.setString(1, invt.getNombre());
-        ps.setString(2, invt.getApellido());
-        ps.setString(3, invt.getDireccion());
-        ps.setString(4, invt.getTelefono());
-        ps.setString(5, invt.getSexo());
-        ps.setString(6, String.valueOf(Conexion.convertirBooleanAChar(invt.isActivo())));
-        ps.setInt(7, invt.getId());
-        ps.executeUpdate();
-        Conexion.desconectar();
+        try {
+            Conexion.conectar();
+            PreparedStatement ps = Conexion.conn.prepareStatement("UPDATE evento SET NOMBRE = ?, FECHA = ?, UBICACION = ?, ACTIVO = ? WHERE ID_EVENTO = ?");
+            ps.setString(1, invt.getNombre());
+            ps.setString(2, invt.getApellido());
+            ps.setString(3, invt.getDireccion());
+            ps.setString(4, invt.getTelefono());
+            ps.setString(5, invt.getSexo());
+            ps.setString(6, String.valueOf(Conexion.convertirBooleanAChar(invt.isActivo())));
+            ps.setInt(7, invt.getId());
+            ps.executeUpdate();
+        } finally {
+            Conexion.desconectar();
+        }
     }
 
     public static void eliminar(int id) throws SQLException {
-        Conexion.conectar();
-        PreparedStatement ps = Conexion.conn.prepareStatement("DELETE FROM invitado WHERE Id_Invitado = ?");
-        ps.setInt(1, id);
-        ps.executeUpdate();
-        Conexion.desconectar();
+        if (id <= 0) {
+            throw new IllegalArgumentException("No se puede Eliminar un registro con id <= 0");
+        }
+        try {
+            Conexion.conectar();
+            PreparedStatement ps = Conexion.conn.prepareStatement("DELETE FROM invitado WHERE Id_Invitado = ?");
+            ps.setInt(1, id);
+            ps.executeUpdate();
+        } finally {
+            Conexion.desconectar();
+        }
     }
 
     public static List<Invitado> SeleccionarPorNombreApellido(String nombre, String apellido) throws SQLException {
@@ -83,7 +94,7 @@ public class InvitadoAcessoDatos {
         PreparedStatement ps = null;
         boolean ejecutar = false;
         ArrayList<Invitado> invitados = null;
-        
+
         if (nombre != null && !nombre.isEmpty() && apellido != null && !apellido.isEmpty()) {
             query = "nombre like ? OR apellido ?";
             ps = Conexion.conn.prepareStatement(query);
@@ -101,7 +112,7 @@ public class InvitadoAcessoDatos {
             ps.setString(1, "%" + apellido + "%");
             ejecutar = true;
         }
-        
+
         if (ejecutar) {
             ResultSet rs = ps.executeQuery();
             invitados = new ArrayList<>();
